requirements.txt
python-docx
docx2pdf
platformdirs
